+++ src/components/inline-tools/inline-tool-convert.ts
@@ -1,6 +1,6 @@
 import { IconReplace } from '@codexteam/icons';
 import { InlineTool, API } from '../../../types';
-import { MenuConfig } from '../../../types/tools';
+import { MenuConfig, MenuConfigItem } from '../../../types/tools';
 import * as _ from '../utils';
 import { Blocks, Selection, Tools, I18n, Caret } from '../../../types/api';
 import SelectionUtils from '../selection';
@@ -57,6 +57,11 @@ export default class ConvertInlineTool implements InlineTool {
   public async render(): Promise<MenuConfig> {
     const currentSelection = SelectionUtils.get();
     const currentBlock = this.blocksAPI.getBlockByElement(currentSelection.anchorNode as HTMLElement);
+
+    if (currentBlock === undefined) {
+      return [];
+    }
+
     const allBlockTools = this.toolsAPI.getBlockTools();
     const convertibleTools = await getConvertibleToolsForBlock(currentBlock, allBlockTools);
 
@@ -64,8 +69,8 @@ export default class ConvertInlineTool implements InlineTool {
       return [];
     }
 
-    const convertToItems = convertibleTools.reduce((result, tool) => {
-      tool.toolbox.forEach((toolboxItem) => {
+    const convertToItems = convertibleTools.reduce<MenuConfigItem[]>((result, tool) => {
+      tool.toolbox?.forEach((toolboxItem) => {
         result.push({
           icon: toolboxItem.icon,
           title: toolboxItem.title,